Refactor Project#status dependencies to Projects::StatusController
  - Move private methods used by `status` action
  - Move specs from `ProjectController#status` to `Projects::StatusController#show`
  - Adapt specs, controllers and necessary views
