+++ src/api/app/views/webui/project/_breadcrumb_items.html.haml
@@ -29,6 +29,3 @@
     - elsif current_page?(project_config_path(@project))
       %li.breadcrumb-item.active{ 'aria-current' => 'page' }
         Configuration
-    - elsif current_page?(project_status_path(@project))
-      %li.breadcrumb-item.active{ 'aria-current' => 'page' }
-        Status
