+++ src/api/app/views/webui/projects/status/_breadcrumb_items.html.haml
@@ -1,0 +1,7 @@
+= render partial: 'webui/main/breadcrumb_items'
+%li.breadcrumb-item
+  = link_to 'Projects', projects_path
+%li.breadcrumb-item
+  = link_to @project, project_show_path(@project)
+%li.breadcrumb-item.active{ 'aria-current' => 'page' }
+  Status
